refactor: HttpServer::upgrade protected method replaced by the use of std::function.
This change aims to provide a behaviour more flexible.

In the previous design, you'd have to duplicate code and play with inheritance
hierachy if you wanna set the same handler to different HttpServer subclasses.
Worse yet, it'd be very difficult to make high-level libraries, because
abstractions developed by different programmers could be incompatible with each
other.